+++ src/fragments/types/TaxTypeFragment.ts
@@ -1,0 +1,13 @@
+/* tslint:disable */
+/* eslint-disable */
+// This file was automatically generated and should not be edited.
+
+// ====================================================
+// GraphQL fragment: TaxTypeFragment
+// ====================================================
+
+export interface TaxTypeFragment {
+  __typename: "TaxType";
+  description: string | null;
+  taxCode: string | null;
+}