+++ src/taxes/queries.ts
@@ -1,11 +1,14 @@
 import {
   countryWithTaxesFragment,
-  shopTaxesFragment
+  shopTaxesFragment,
+  taxTypeFragment
 } from "@saleor/fragments/taxes";
+import makeQuery from "@saleor/hooks/makeQuery";
 import gql from "graphql-tag";
 
 import { TypedQuery } from "../queries";
 import { CountryList } from "./types/CountryList";
+import { TaxTypeList } from "./types/TaxTypeList";
 
 const countryList = gql`
   ${countryWithTaxesFragment}
@@ -20,3 +23,13 @@ const countryList = gql`
   }
 `;
 export const TypedCountryListQuery = TypedQuery<CountryList, {}>(countryList);
+
+const taxTypeList = gql`
+  ${taxTypeFragment}
+  query TaxTypeList {
+    taxTypes {
+      ...TaxTypeFragment
+    }
+  }
+`;
+export const useTaxTypeList = makeQuery<TaxTypeList, {}>(taxTypeList);